Update operator-sdk to 1.22.1
In doing so, this patch also goes back to 1.15.0 (which I think is
where we started), and applies all of the migrations that are specified
at https://sdk.operatorframework.io/docs/upgrading-sdk-version/
It should be noted that in the future, this should be check every
time operator-sdk is updated here.

This also removes the bundle/ dir, and only keeps the dependencies.yaml.

Signed-off-by: Brad P. Crochet <brad@redhat.com>